Adds Spine Runtime and Spine components, Convict asset with animations.
I imported the spine runtime so that I could get the convict asset into the project. Added a new script to handle animation for convict character. The spine animation component is now attached to the Unit object, and I made the Unit object into a prefab. Also, for funsies, I added a fuckton of convicts into the scene just to see what would happen. Was not disappointed.